+++ AStarProblems/StandardMultiAgentAStarProblemWithConstraints.h
@@ -44,8 +44,8 @@ private:
     // prePositions is the positions of the agents in state
     // t is the timestep in state (when we add a successor to the list of successors, its timestep is t+1)
     // cost and violations are the cost and the number of violated soft constraints by assigning the agents from 0 to agentToAssign
-    // cannotMove is the list of agents which are at their target positions and cannot move anymore (for the SumOfCosts objective function)
-    void recursiveAssignAMoveToAnAgent(int agentToAssign, std::vector<std::tuple<std::shared_ptr<StandardMultiAgentSpaceTimeState>, int, int>>* successors, int cost, std::vector<int> positions, const std::vector<int>& prePositions, int t, int violations, std::vector<int> cannotMove = std::vector<int>()) const ;
+    // cannotMove[i] is true if agent i is at its target position and cannot move anymore (for the SumOfCosts objective function)
+    void recursiveAssignAMoveToAnAgent(int agentToAssign, std::vector<std::tuple<std::shared_ptr<StandardMultiAgentSpaceTimeState>, int, int>>* successors, int cost, std::vector<int> positions, const std::vector<int>& prePositions, int t, int violations, std::vector<u_int8_t> cannotMove = std::vector<u_int8_t>()) const ;
 
 };
